requirements.txt
starlette
aiohttp
python-dotenv
jsonpath-ng
sentry-sdk
